requirements.txt
black==24.10.0
click==8.1.7
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
python-magic==0.4.27
tqdm==4.66.5
